Remove ${PN} from SRC_URI
In newer yocto versions it seems that the ${PN} variable is not
expanded anymore if it is in the SRC_URI, we got the error:

-------------------
ERROR: ExpansionError during parsing
meta-bistro/layers/multimedia-layer/recipes-multimedia/dleyna/dleyna-core_git.bb:
Failure expanding variable SRCPV, expression was ${@bb.fetch2.get_srcrev(d)}
which triggered exception FetchError: Fetcher failure: The
SRCREV_FORMAT variable must be set when multiple SCMs are used.
-------------------

Replacing the ${PN} with ${BPN} makes it possible to use this layer again.