distro: Remove bistro as a Yocto distribution
Remove references to bistro as a Yocto distribution. Remove image recipes
from the meta-bistro layer.

Bistro is being phased out as a Yocto distribution and is being replaced
by the PELUX distribution that is based on poky. The maintenance burden
and divergence from poky were the main reasons for phasing out bistro.

The PELUX distribution can be found as a part of the meta-pelux layer in
the following repository: https://github.com/Pelagicore/meta-pelux

As a layer meta-bistro will continue to exist as a component layer that
provides IVI related functionality and recipes for Pelagicore sofware
components.

Signed-off-by: Gordan Markuš <gordan.markus@pelagicore.com>