gdbus-codegen-glibmm: Fix build issue for hard-coded version
The sed script in the inc file had a hard-coded version number, but if
we would like this recipe to be more future proof and be able to handle
more than one version of the tool, then we need to fetch the version nr
from somewhere.

The solution is to name the bb file after the major version of the tool
and fetch the number in the inc file.

Signed-off-by: Therese Nordqvist <therese.nordqvist@pelagicore.com>