connectivity-manager: add recipe
Add a runtime dependency on ConnMan. Only backend implemented. In the
future we should perhaps autodetect (and auto configure)
connectivity-manager depending on what "network manager" is used.

Do not need a build dependency on ConnMan since it does not include a
D-Bus interface file (a custom one exists in the connectivity-manager
repository).

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>