user-identification-manager: update to rev with smart card id source
And make it possible to enable/disable through PACKAGECONFIG.

The pcsc-lite daemon ships with systemd socket activation by default so no
extra configuration is needed. It will be started automatically when
user-identification-manager tries to comminicate with it through the
libpcsclite API.

Add ccid as a runtime dependency as well as pcsc-lite. pcsc-lite does not
ship with any drivers by default. May reconsider ccid as a runtime
dependency at some point and require it to be added at a "higher level".

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>